Updated to a basic visual feedback via flashing -> Added epilepsy trigger warning -> Added a bunch of draw primitives to the WindowHandle class -> Added a VisualOutput class for handling the visualizations -> Added data structures for some basic primitives
Currently only the flash visualization is working; metronome and raindrops are not yet implemented (they will be added in a different branch).  There appears to be a BPM limit based on the frequency at which draws can occur, probably the upper limit is around 300BPM or something, though I don't imagine people wanting to go much higher anyway.
-> Need to improve colour management method
-> Need to document the new methods and classes which have been added
-> Need to profile performance before merge to main branch